Add files via upload
In this project, I used the Auto Machine Learning TPOT. It is an open-source framework for automatic machine learning. It uses the genetic engineering to incorporate the machine learning tasks. It is a very efficient machine-learning framework. In this project, I used the Breast Cancer dataset from the sklearn library.