Add files via upload
In this project, I used the Auto Machine Learning Pychart.  It is an open-source framework for automatic machine learning. It uses the genetic engineering to incorporate the machine learning tasks. It is mostly used in automated ML frameworks among Data Science and AI communities.  It is a very efficient machine-learning framework. This project takes the dataset of Diabetes from the Pycaret library. In addition, it provides a comprehensive list of hyperparameter tuning options to male the prediction accurate.